fast_api/requirements.txt
# API + SERVER
fastapi==0.104.1
uvicorn[standard]==0.24.0
# BANCO DE DADOS
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
# TRATAMENTO DE DADOS
mne==1.6.0
# CONFIGS
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
numpy==1.26.2
pandas>=2.1.3
# TOKENS JWT
python-jose[cryptography]==3.3.0
# HASHING DE SENHAS 
bcrypt==4.0.1
passlib==1.7.4
# CLASSIFICATION
scikit-learn==1.3.2
scipy==1.11.3
